ats_resume_analyzer/requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
